+++ src/diag.h
@@ -103,6 +103,7 @@ void diag_emit(diag_t *d);
 void diag_femit(diag_t *d, FILE *f);
 void diag_suppress(diag_t *d, bool suppress);
 void diag_stacktrace(diag_t *d, bool stacktrace);
+void diag_clear(diag_t *d);
 
 unsigned error_count(void);
 void reset_error_count(void);